# 01-todo-list-api/openapi.json
{
  "openapi": "3.0.3",
  "info": { "title": "To-do List API", "version": "1.0.0" },
  "servers": [{ "url": "http://localhost:3000" }],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/todos": {
      "get": {
        "summary": "Listar todos",
        "responses": {
          "200": {
            "description": "Lista de todos",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Todo" } } } }
          }
        }
      },
      "post": {
        "summary": "Criar todo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateTodo" } }
          }
        },
        "responses": { "201": { "description": "Criado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Todo" } } } } }
      }
    },
    "/todos/{id}": {
      "patch": {
        "summary": "Atualizar todo",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateTodo" } } }
        },
        "responses": { "200": { "description": "Atualizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Todo" } } } },
          "404": { "description": "Não encontrado" } }
      },
      "delete": {
        "summary": "Apagar todo",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Apagado" }, "404": { "description": "Não encontrado" } }
      }
    }
  },
  "components": {
    "schemas": {
      "Todo": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "title": { "type": "string" }, "done": { "type": "boolean" } },
        "required": ["id", "title", "done"]
      },
      "CreateTodo": {
        "type": "object",
        "properties": { "title": { "type": "string" }, "done": { "type": "boolean" } },
        "required": ["title"]
      },
      "UpdateTodo": {
        "type": "object",
        "properties": { "title": { "type": "string" }, "done": { "type": "boolean" } }
      }
    }
  }
}